Cambio de nombre a rabbitmq_client y declaración de tipo Callable
